Split up the Arrow monolith
The old vtab/arrow.rs had grown to 3,212 lines, with shared Arrow
schema and vector conversion code mixed into the table-function adapter.

Move that shared conversion code into a private arrow_interop module
while keeping existing vtab re-exports intact. Leave vtab::arrow focused
on the table-function adapter and query parameter glue.
